switch from int to float for AWS preparations (#13)

+++ service/creator/request/memory.go
@@ -2,7 +2,7 @@ package request
 
 // Memory configures the machine memory.
 type Memory struct {
-	SizeGB int `json:"size_gb"`
+	SizeGB float64 `json:"size_gb"`
 }
 
 // DefaultMemory provides a default ram configuration by best effort.
